added getUserV2 api method
