+++ services/app-api/serverless.yml
@@ -14,11 +14,11 @@ plugins:
   - serverless-dynamodb-local
   - serverless-associate-waf
   - serverless-offline
-  - serverless-iam-helper
-  - serverless-s3-bucket-helper
+  # - serverless-iam-helper
+  # - serverless-s3-bucket-helper
 custom:
   stage: ${opt:stage, self:provider.stage}
-  oneMacTableName: onemac-${self:custom.stage}-one
+  oneMacTableName: onemac-develop-one
   iamPath: ${ssm:/configuration/${self:custom.stage}/iam/path, ssm:/configuration/default/iam/path, "/"}
   iamPermissionsBoundaryPolicy: ${ssm:/configuration/${self:custom.stage}/iam/permissionsBoundaryPolicy, ssm:/configuration/default/iam/permissionsBoundaryPolicy, ""}
   emailSource: ${ssm:/configuration/${self:custom.stage}/email/cms_spa_form_from_email, ssm:/configuration/default/email/cms_spa_form_from_email, file(resources/ssm-params.yml):Resources.CmsSpaFormFromEmail.Properties.Value}
@@ -567,6 +567,16 @@ functions:
     events:
       - http:
           path: getUser
+          method: get
+          cors: true
+          authorizer: aws_iam
+
+  getUserV2:
+    handler: getUserV2.main
+    role: LambdaApiRole
+    events:
+      - http:
+          path: getUserV2
           method: post
           cors: true
           authorizer: aws_iam
